+++ studips/src/App.css
@@ -25,10 +25,18 @@
 	align-items: center;
 }
 
-#category {
+.category {
 	align-self: flex-end;
 	color: #a40032;
 	margin-left: 3vw;
+	background: none;
+	border: none;
+	outline: none;
+	text-decoration: underline;
+	cursor: pointer;
+}
+.category:hover {
+	text-decoration: none;
 }
 
 .avatar {
@@ -67,24 +75,29 @@
 	font-size: 12px;
 	align-self: flex-end;
 }
-.triangle {
+.message:before {
+	position: absolute;
+	content: '';
+	display: block;
     width: 0; 
     height: 0; 
     border-left: 10px solid transparent;
     border-right: 10px solid transparent;    
     border-bottom: 10px solid #a40032;
     align-self: flex-start;
-    margin-left: 2em;
-    z-index: 10;
+	z-index: 10;
+	top: -10px;
+    left: 3em;
 }
 .message {
-	width: 95%;
+	position: relative;
 	align-items: center;
-	padding: 1%;
+	padding: .5rem 1rem;
 	background-color: #a40032;
 	color: white;
 	font-size: 12px;
 	box-shadow: 0px 2px 4px black;
+	margin-bottom: .5rem;
 }
 .actions {
 	width: 100%;
