+++ studips/src/App.js
@@ -1,20 +1,29 @@
-import React from 'react';
+import React, { useState } from 'react';
 import MainThread from './component/MainThread';
 import './App.css';
-import ConnectForm from './component/ConnectForm'
-import LoginForm from './component/LoginForm.js'
+import LoginForm from './component/LoginForm'
+import RegisterForm from './component/RegisterForm'
 import Welcome from './component/Welcome'
 import {Switch, Route} from 'react-router-dom'
 
 
 function App() {
+	const [token, setToken] = useState(null);
 	return (
 		<div className='App'>
 			<Switch>
-				<Route exact path ='/' component={Welcome} />
-				<Route path = '/connect' component={ConnectForm} />
-				<Route path='/login' component={LoginForm} />
-				<Route path ='/mainthread' component={MainThread} />
+				<Route exact path ='/'>
+					<Welcome />
+				</Route>
+				<Route path = '/login'>
+					<LoginForm token={token} setToken={setToken} />
+				</Route>
+				<Route path='/register'>
+					<RegisterForm />
+				</Route>
+				<Route path ='/mainthread'>
+					<MainThread token={token} />
+				</Route>
       		</Switch>
 		</div>
 	);
